Added unit tests for remote weavelet. (#528)
Added unit tests for remote weavelet.

This PR adds unit tests for the remote weavelet. Specifically, it tests
that the weavelet still operates correctly even when some of its
interactions with the envelope fail. For example, the weavelet should
retry failed ActivateComponent calls. A handful of the tests are
currently failing because of "bugs" in the remote weavelet. I skipped
the tests for now and will try to fix the bugs in future PRs.

Fixing the bugs is a bit tricky because it's not always straightforward
the best way to handle failures. For example, if a weavelet calls
GetListenerAddress and receives an address that is already in use,
should it keep retrying the GetListenerAddress? Or should it just
self-terminate?