Verify registered listeners when starting the app.
This ensures that the application listeners match the information
stored in the read-only section of the application binary, and used
by various deployers. If there is a mismatch, the user will be prompted
to run `weaver generate`.

This fixes one of the common issues that came up in workshops, where
people forgot to re-run `weaver generate` after declaring the listener.