Added basic tracing to the simulator. (#566)
Recall that a simulation consists of a number of discrete events:

- An op starts.
- An op finishes.
- A method call is made.
- A method call is delivered.
- A method call returns.
- A method return return is delivered.
- An error is injected.

This PR adds basic tracing to the simulator to keep track of these
events. When a simulation fails, a history of the events is returned to
the user. They can inspect the history to understand why the simulation
failed. In a future PR, I'll add a way to visualize a history of events.

In the future, traces could be used for more than just visualization.
For example, we may use traces for minimization or for checking that a
simulation is deterministic. Thus, I expect the contents of the traces
to change over time as we refine the simulator. I went with something
simple for now.