Added hyperparameter for simulator failure rate. (#569)
Recall that the simulator can arbitrarily choose to fail a component
method call, either before or after delivering the call. Before this PR,
the simulator had an incredibly naive algorithm for deciding when to
fail calls. This naive approach failed a lot of calls.

This PR introduces a hyperparameter `FailureRate` to configure how often
the simulator fails calls. When FailureRate is 0, the simulator never
fails a call, and when FailureRate is 1, the simulator fails every call.