Added simulator yield rate to tune interleaving. (#572)
This PR introduces a new hyperparameter to the simulator, *yield rate*,
that dictates how much an execution is interleaved. Recall that a
simulation consists of a number of operations (ops), and every op
executes a series of steps, mostly calling component methods.

After executing a step, an op makes a decision to keep executing or to
yield control back to the scheduler. The yield rate determines how often
the op yields. With a yield rate of 0, the op never yields. Ops are run
to completion, and there is no interleaving. With a yield rate of 1, ops
yield after every step, leading to a very high degree of interleaving.
With a yield rate of 1/n, an op yields after n steps in expectation. In
general, yielding is governed by a [geometric distribution][geometric].

Note that some of the code in this PR is really inefficient. I left it
as is to keep thing easy to understand, but will optimize things in a
future PR.

[geometric]: https://en.wikipedia.org/wiki/Geometric_distribution