Implemented bare bones simulation with new API.
This PR hooks up the new simulation API to the existing simulation
implementation code. This mostly consists of massaging a workload struct
into types that the underlying simulation engine understands. For now,
the simulator only runs a single simulation on a fixed set of
hyperparameters. In the future, it will sweep over hyperparameters given
a time budget.