Sweep over hyperparameters, running multiple sims.
Recall the `Simulator.Run` method:

```go
func (s *Simulator) Run(duration time.Duration) Results
```

Before this PR, `Run` ran a single simulation. This PR enhances the
simulator to continuously run simulations for the provided duration.

Recall that the behavior of a simulation is governed by a number of
hyperparameters like the RNG seed, number of operations to run, the
probability of injecting a failure, and so on. Given a time budget to
run simulations, how do we explore this hyperparameter space?

That's a *very* hard and open ended question. This PR implements a very
stupid yet simple algorithm that tries to search "simpler" simulations
before exploring more complex simulations. This is only a temporary
algorithm to get things working. In the future, we'll have to revisit
and improve it. The algorithm will also get more complex when we
implement minimization.

This PR also changes the signature of `Run` to not return an error. A
simulation itself should never fail to execute properly.