Use the Go BuildInfo to read the tool and application module versions.
This remove the manual module version tracking (in the version.go file),
which will hopefully lead to fewer versioning errors and allow us to
track the version information at the finer granularity.

Note that the tool and the application module versions aren't ever
directly compared: the output is only used for nicer error message
generation.