Catch panics in simulations.
This PR extends the simulator to catch any panics that happen during an
execution and abort the execution. By catching panics, the simulator is
able to show the user the history of events that led to the panic.

This PR also fixes a minor bug. Previously, we recorded the history for
aborted events. With this PR, once an execution is aborted, no other
events are recorded.

I also renamed some unit tests to be more descriptive.