Detect bad context propagation in simulations.
An executor embeds trace and span ids in the contexts passed to a
workload. The workload is responsible for propagating these contexts
throughout the workload. If a context is not propagated, then an
executor cannot tell which op a given call is part of. This makes it
impossible for the executor to schedule things correctly. Plus, the
reported history will be missing tracing information, making it much
harder to understand what happened.

This PR detects and panics when contexts are not being propagated
properly. I think with enough effort, we could modify the executor to
gracefully handle missing trace information, but I'll leave that for
future work.