Added instructions to run chat app on Kubernetes.
This PR adds instructions on how to run the chat app on a Kubernetes
cluster using the `weaver kube` deployer. The main challenge is running
MySQL in the Kubernetes cluster. It is quite a bit more involved than
running MySQL locally using Docker. You have to write about 100 lines of
Kubernetes YAML, which isn't rocket science, but it also isn't the most
pleasant experience.