Exported Resolver interface in run package.
I plan on moving the `run` package to the `runtime/` directory to allow
other deployers, like `weaver kube`, to use it. However, the `run`
package uses the `Resolver` interface, which is currently in the
internal `call` package. This PR copies the `Resolver` interface into
the `run` package and also simplifies it a bit (replacing endpoints with
addresses).