Added newer, hopefully simpler, Resolver API.
Recall that `run.Run` runs a weavelet with a fixed set of components.
When this weavelet wants to call a method on a remote component, it
needs to know the addresses of the weavelets hosting the remote
component. To get these addresses, `run.Run` uses a set of provided
resolvers.

>> Existing Resolver API

Previously, the `run` package used the `call.Resolver` interface to
represent resolvers:

```go
type Resolver interface {
    IsConstant() bool
    Resolve(ctx context.Context, version *Version) ([]Endpoint, *Version, error)
}

type Version struct {
    Opaque string
}
```

We designed this API according to the principle of "preferring
synchronous functions", as detailed in Bryan Mill's GopherCon 2018 talk
[Rethinking Classical Concurrency Patterns][sync]. In short, your API
should include synchronous function calls whenever possible. This avoids
some ambiguities around goroutine lifetimes, context cancellations,
channel buffering, channel closing, etc. Plus, you can easily introduce
asynchrony to a synchronous function (just spawn a goroutine), but you
cannot as easily remove asynchrony.

The existing `Resolver` API has a lot of advantages, but, the more I
used it, the more I started to worry that our decision to adhere to the
"prefer synchronous functions" principle was more *dogmatic* that
*pragmatic*. Here are some disadvantages of the API:

1. It is complicated. Understanding versions, in particular, requires a
   fair bit of effort. `Resolve` sometimes blocks and sometimes doesn't,
   depending on the provided version. `Resolve` returns a nil version
   when the resolver is constant, a new version when the resolver is not
   constant, and the same version when the resolution times out.
   Understanding these nuances makes using and writing resolvers hard.

2. It doesn't compose well. In writing the new `weaver exec` command, I
   needed to write a `Resolver` that combined the addresses from a set
   of other `Resolver`s. This proved to be very complicated. For
   example, you have to invent a new set of versions for the combined
   resolver. And because a resolver has to be thread-safe, the logic
   around waking up blocked calls to `Resolve` gets finicky. It's
   doable, but feels unnecessarily complicated.

3. Inherent asynchrony. Even though the `Resolver` API is synchronous,
   many resolvers (e.g., a DNS resolver, a Kubernetes resolver, a
   ConfigMap resolver) are inherently doing work in the background. They
   don't only resolve when you call `Resolve`. They are constantly
   resolving in the background. This means that many of the resolvers
   have the issues of context cancellation and goroutine lifetimes that
   synchronous calls try to avoid.

4. Versions are not typed. All versions have to be converted into a
   string. Moreover, the string shouldn't be used by users, yet it must
   be exported to allow people to implement resolvers.

>> New Resolver API

This PR introduces a new `Resolver` API.

```go
type Resolver interface {
    Addresses() <-chan []string
    Err() error
}
```

A `Resolver` returns a channel of addresses. If the resolver encounters
an error, it closes the channel, and the `Err` method returns the error.

The API is a lot simpler due to the lack of versions. It also composes
much better; see the `CombinedResolver` implementation in this PR.

The API also has some drawbacks:

- If you don't read from an `Addresses` channel, you don't know when the
  writing end of the channel will block. And when implementing a
  `Resolver`, you have to be prepared for your writes to block.
- If a resolver encounters an error, but there are a number of addresses
  buffered in the channel, you have to read all the addresses before
  realizing the channel is closed.

[sync]: https://www.youtube.com/watch?v=5zXAHh5tJqQ