Add optional names for colocation groups
We've discussed to enable the user to specify group names.
Right now we compute automatically the group names as the name of the
first component in each group.

This PR enables the user to specify group names (optionally). This is
becoming more and more important given that we allow the user to
configure various knobs in the deployer, and many knobs make sense for a
particular colocation group. E.g., secrets/config maps make sense for colocation
group. Maybe we will allow people to configure resource requirements per
colocation group at some point, etc.