Colocation groups config
Right now we specify the components that should run together in a
colocation group. For all the others we run them in their own groups.
However, given that the group is the unit of deployment, we'll have to
add configuration knobs that make sense per colocation group.

This PR provides some options to set per colocation group knobs in the
config.