Fix MTLS handling. (#707)
We were using the wrong component name when setting up group data
structures for MTLS (because of Go's funny loop variable capture
semantics). Fixed this.

Added a test that uses multi deployer with MTLS (disabled on Github due to its slowness).

Increased timeout for the test from 4s to 10s.