Remove envelope<->weavelet pipes. (#712)
Initialization info sent by the envelope to the weavelet is now sent
in an environment variable.

Initialization info sent by the weavelet to the envelope is now sent
as the response to a control.WeaveletControl.InitWeavelet() call. This
call is made by NewEnvelope() as an initial hand-shake right after
creating the weavelet.

Deleted all the envelope/weavlet connection management code.

Deleted low-level tests that were testing details of the connection
management protocol.

Dropped unnecessary code that carried a Bootstrap object in a context.
We had switched away from our use of a Bootstrap context value a long
time ago.