Exported the `sim` package.
This PR moves the `sim` package out of `internal/` into the root of the
weaver repository, exporting it for public use.

We may want to merge the `sim` package into the `weavertest` package.
Or, we could keep `sim` as a standalone package, maybe renaming it to
`weaversim`. I'm deferring either of these decisions for now.