Add generated label to autogenerated metrics
Service Weaver users complain that we're exporting too many metrics,
hence the cloud bill is high. This is mainly because we export a bunch
of metrics that are either autogenerated or created within our system
components.

This PR adds a new label to each of these metrics, s.t. the deployer can
decide to filter them out, before exporting the metrics for consumption.