Add shutdown method per component (#766)
In the current implementation, we provide the user the ability to
initialize things by implementing an `Init` method. However, they don't
have the ability to do a clean shutdown easily.

The user filled multiple github issues (#275 and #765) asking for a `Shutdown`
method.

This PR adds a `Shutdown` method to each component. Note that there is
no guarantee that this method can run - e.g., the app receives a
SIGKILL because the machine suddenly crashes. However, a `Shutdown`
method can enable the user to achieve graceful shutdown in normal
scenarios.

In tests, if someone wants to test the `Shutdown` method, they can get a
pointer to the implementation of the component and explicitly call the
`Shutdown` method.