Remove NonNegativeBuckets
NonNegativeBuckets were used for our generated metrics. However, it is
in the public metrics package, and users can rely on it to define their
histograms.

However, NonNegativeBuckets has a huge number of buckets which lead to
large data for the metrics. The users have complained that our buckets are
too fine-grained and the cost to export the metrics is quite high.

This PR removes the NonNegativeBuckets and forces the user to define
their own buckets for their histograms.