Fix traffic graph
After we introduced system components (weavelet control and deployment
control), the traffic graph doesn't display anything in the weaver multi
dashboard.

This is because when we display the graph, we look at the components
of the application and then we try to match them to the traffic between
them. However, the edges show traffic for components that are missing from the
status (e.g., the system components). Hence, when we display the graph
there is no 1:1 mapping between components and the traffic between
them.

This PR fixes this issue by not displaying traffic for system
components (in fact the user shouldn't care about that). Maybe a better
fix would be to do this in the javascript itself, but we can do that in
the future if needed.